Homework 6 completed

+++ core-service/src/main/resources/db/migration/V1__init.sql
@@ -12,7 +12,7 @@ create table products
     id              bigserial primary key,
     title           varchar(255),
     category_id     bigint references categories (id),
-    price           int,
+    price           decimal,
     created_at      timestamp default current_timestamp,
     updated_at      timestamp default current_timestamp
 );
@@ -24,7 +24,7 @@ create table orders
 (
     id          bigserial primary key,
     username    varchar(255) not null,
-    total_price int not null,
+    total_price decimal not null,
     address     varchar(255),
     phone       varchar(255),
     created_at  timestamp default current_timestamp,
@@ -37,8 +37,8 @@ create table order_items
     product_id        bigint not null references products (id),
     order_id          bigint not null references orders (id),
     quantity          int    not null,
-    price_per_product int    not null,
-    price             int    not null,
+    price_per_product decimal   not null,
+    price             decimal   not null,
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
 );
